Extract applyImageFit logic.
Also, add FractionalOffset.inscribe.
