+++ packages/flutter/lib/painting.dart
@@ -21,7 +21,9 @@ export 'src/painting/basic_types.dart';
 export 'src/painting/box_painter.dart';
 export 'src/painting/colors.dart';
 export 'src/painting/decoration.dart';
+export 'src/painting/image_fit.dart';
 export 'src/painting/edge_insets.dart';
+export 'src/painting/fractional_offset.dart';
 export 'src/painting/text_editing.dart';
 export 'src/painting/text_painter.dart';
 export 'src/painting/text_span.dart';
